backend/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
pydantic>=2.0.0
